# common-platform-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/organization'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/banks/{bankId}/divisions/{divisionId}/regions/{regionId}':
    get:
      tags:
        - Bank Region Management
      summary: Get bank region by ID
      description: Returns a region of a specific bank division based on its ID
      operationId: getBankRegionById
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: divisionId
          in: path
          description: ID of the division
          required: true
          schema:
            type: integer
            format: int64
        - name: regionId
          in: path
          description: ID of the region to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved bank region
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '404':
          description: 'Bank, division, or region not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
    put:
      tags:
        - Bank Region Management
      summary: Update bank region
      description: >-
        Updates an existing region of a specific bank division with the provided
        details
      operationId: updateBankRegion
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: divisionId
          in: path
          description: ID of the division
          required: true
          schema:
            type: integer
            format: int64
        - name: regionId
          in: path
          description: ID of the region to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankRegionDTO'
        required: true
      responses:
        '200':
          description: Bank region successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '400':
          description: Invalid bank region data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '404':
          description: 'Bank, division, or region not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
    delete:
      tags:
        - Bank Region Management
      summary: Delete bank region
      description: Deletes a region of a specific bank division based on its ID
      operationId: deleteBankRegion
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: divisionId
          in: path
          description: ID of the division
          required: true
          schema:
            type: integer
            format: int64
        - name: regionId
          in: path
          description: ID of the region to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Bank region successfully deleted
        '404':
          description: 'Bank, division, or region not found'
        '500':
          description: Internal server error
...
  '/api/v1/banks/{bankId}':
    get:
      tags:
        - Bank Management
      summary: Get bank by ID
      description: Returns a bank based on its ID
      operationId: getBankById
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved bank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
    put:
      tags:
        - Bank Management
      summary: Update bank
      description: Updates an existing bank with the provided details
      operationId: updateBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankDTO'
        required: true
      responses:
        '200':
          description: Bank successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '400':
          description: Invalid bank data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankDTO'
    delete:
      tags:
        - Bank Management
      summary: Delete bank
      description: Deletes a bank based on its ID
      operationId: deleteBank
      parameters:
        - name: bankId
          in: path
          description: ID of the bank to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Bank successfully deleted
        '404':
          description: Bank not found
        '500':
          description: Internal server error
...
  '/api/v1/banks/{bankId}/divisions/{divisionId}/regions/filter':
    post:
      tags:
        - Bank Region Management
      summary: Get all regions for a bank division with filtering
      description: >-
        Returns a paginated list of regions for a specific bank division based
        on filter criteria
      operationId: filterBankRegions
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: divisionId
          in: path
          description: ID of the division
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBankRegionDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved bank regions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankRegionDTO'
        '404':
          description: Bank or division not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankRegionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankRegionDTO'
  '/api/v1/banks/{bankId}/divisions/{divisionId}/regions':
    post:
      tags:
        - Bank Region Management
      summary: Create a new region for a bank division
      description: >-
        Creates a new region for a specific bank division with the provided
        details
      operationId: createBankRegion
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: divisionId
          in: path
          description: ID of the division
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankRegionDTO'
        required: true
      responses:
        '201':
          description: Bank region successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '400':
          description: Invalid bank region data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '404':
          description: Bank or division not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankRegionDTO'
...
  /api/v1/banks/filter:
    post:
      tags:
        - Bank Management
      summary: Get all banks with filtering
      description: Returns a paginated list of banks based on filter criteria
      operationId: filterBanks
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBankDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved banks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBankDTO'
...
components:
  schemas:
...
    BankRegionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        divisionId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    BankDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        websiteUrl:
          type: string
        addressLine:
          type: string
        postalCode:
          type: string
        city:
          type: string
        state:
          type: string
        countryId:
          type: integer
          format: int64
        timeZoneId:
          type: integer
          format: int64
        isActive:
          type: boolean
        establishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestBankRegionDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BankRegionDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for bank regions
    PaginationResponseBankRegionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BankRegionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestBankDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BankDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for banks
    PaginationResponseBankDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BankDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.